requirements.txt
numpy
pandas
matplotlib
seaborn
scipy
scikit-learn
pyspark
xgboost
jupyterlab
apache-airflow==2.8.1
tqdm

# ML Metrics and Model Tracking
mlflow>=2.8.0
psycopg2-binary>=2.9.0
hyperopt>=0.2.7

# Monitoring and Metrics
prometheus-client>=0.17.0
statsd>=4.0.1

# Additional ML Libraries
lightgbm>=4.0.0

# Data Validation
great-expectations>=0.17.0

# Visualization
plotly>=5.15.0
dash>=2.14.0